Moved StateHandler Moved funcs from unified in to StateHandler WIP - StateHandler changes to fix smartMode issue AC humidity fixes Tidied up / unified accessory definition files
