+++ homekit/AirPurifier.js
@@ -23,17 +23,14 @@ class AirPurifier {
 		this.roomName = deviceInfo.roomName
 		this.name = this.roomName + ' Pure'
 		this.type = 'AirPurifier'
-		this.displayName = this.name
 		this.disableLightSwitch = platform.disableLightSwitch
 		this.filterService = deviceInfo.filterService
 		this.capabilities = unified.capabilities(device, platform)
 
+		const StateHandler = require('./StateHandler')(this, platform)
+
 		this.state = this.cachedState.devices[this.id] = unified.acState(device)
-
-		const StateHandler = require('../sensibo/StateHandler')(this, platform)
-
 		this.state = new Proxy(this.state, StateHandler)
-
 		this.stateManager = require('./StateManager')(this, platform)
 
 		this.UUID = this.api.hap.uuid.generate(this.id)
@@ -48,6 +45,7 @@ class AirPurifier {
 			this.accessory.context.deviceId = this.id
 
 			platform.cachedAccessories.push(this.accessory)
+
 			// register the accessory
 			this.api.registerPlatformAccessories(platform.PLUGIN_NAME, platform.PLATFORM_NAME, [this.accessory])
 		}
