+++ sensibo/unified.js
@@ -13,34 +13,6 @@ function fanLevelToHK(value, fanLevels) {
 	return Math.round(100 * valueIndex / totalLevels)
 }
 
-function HKToFanLevel(value, fanLevels) {
-	let selected = 'auto'
-
-	if (!fanLevels.includes('auto')) {
-		selected = fanLevels[0]
-	}
-
-	if (value !== 0) {
-		fanLevels = fanLevels.filter(level => {
-			return level !== 'auto'
-		})
-		const totalLevels = fanLevels.length
-
-		for (let i = 0; i < fanLevels.length; i++) {
-			if (value <= Math.round(100 * (i + 1) / totalLevels)) {
-				selected = fanLevels[i]
-				break
-			}
-		}
-	}
-
-	return selected
-}
-
-function toFahrenheit(value) {
-	return Math.round((value * 1.8) + 32)
-}
-
 function toCelsius(value) {
 	return (value - 32) / 1.8
 }
@@ -252,104 +224,6 @@ module.exports = {
 		const state = { occupancy: (location.occupancy === 'me' || location.occupancy === 'someone') ? 'OCCUPANCY_DETECTED' : 'OCCUPANCY_NOT_DETECTED' }
 
 		return state
-	},
-
-	sensiboFormattedACState: (device, state) => {
-		device.log.easyDebug(`${device.name} -> sensiboFormattedState: ${JSON.stringify(state, null, 4)}`)
-		const acState = {
-			on: state.active,
-			mode: state.mode.toLowerCase(),
-			temperatureUnit: device.temperatureUnit,
-			targetTemperature: device.usesFahrenheit ? toFahrenheit(state.targetTemperature) : state.targetTemperature
-		}
-
-		// TODO: this is duplicated in sensiboFormattedClimateReactState below, could it be a helper function instead?
-		if ('threeDimensionalSwing' in device.capabilities[state.mode]) {
-			if ((state.horizontalSwing === 'SWING_ENABLED') && (state.verticalSwing === 'SWING_ENABLED')) {
-				acState.swing = 'both'
-			} else if (state.verticalSwing === 'SWING_ENABLED') {
-				acState.swing =  'rangeFull'
-			} else if (state.horizontalSwing === 'SWING_ENABLED') {
-				acState.swing = 'horizontal'
-			} else {
-				acState.swing = 'stopped'
-			}
-		} else {
-			if ('verticalSwing' in device.capabilities[state.mode]) {
-				acState.swing = state.verticalSwing === 'SWING_ENABLED' ? 'rangeFull' : 'stopped'
-			}
-
-			if ('horizontalSwing' in device.capabilities[state.mode]) {
-				acState.horizontalSwing = state.horizontalSwing === 'SWING_ENABLED' ? 'rangeFull' : 'stopped'
-			}
-		}
-
-		if ('fanSpeeds' in device.capabilities[state.mode]) {
-			acState.fanLevel = HKToFanLevel(state.fanSpeed, device.capabilities[state.mode].fanSpeeds)
-		}
-
-		if ('light' in device.capabilities[state.mode]) {
-			acState.light = state.light ? 'on' : 'off'
-		}
-
-		return acState
-	},
-
-	sensiboFormattedClimateReactState: (device, state) => {
-		const smartModeState = state.smartMode
-		const climateReactState = {
-			enabled: smartModeState.enabled,
-			type: smartModeState.type,
-			lowTemperatureThreshold: smartModeState.lowTemperatureThreshold,
-			lowTemperatureState:
-			{
-				on: smartModeState.lowTemperatureState.on,
-				light: smartModeState.lowTemperatureState.light ? 'on' : 'off',
-				temperatureUnit: device.temperatureUnit,
-				fanLevel: HKToFanLevel(state.fanSpeed, device.capabilities[state.mode].fanSpeeds),
-				mode: smartModeState.lowTemperatureState.mode.toLowerCase(),
-				targetTemperature: smartModeState.lowTemperatureState.targetTemperature
-			},
-			lowTemperatureWebhook: null,
-			highTemperatureThreshold: smartModeState.highTemperatureThreshold,
-			highTemperatureState:
-			{
-				on: smartModeState.highTemperatureState.on,
-				light: smartModeState.highTemperatureState.light ? 'on' : 'off',
-				temperatureUnit: device.temperatureUnit,
-				fanLevel: HKToFanLevel(state.fanSpeed, device.capabilities[state.mode].fanSpeeds),
-				mode: smartModeState.highTemperatureState.mode.toLowerCase(),
-				targetTemperature: smartModeState.highTemperatureState.targetTemperature
-			},
-			highTemperatureWebhook: null
-		}
-
-		if ('threeDimensionalSwing' in device.capabilities[state.mode]) {
-			if ((state.horizontalSwing === 'SWING_ENABLED') && (state.verticalSwing === 'SWING_ENABLED')) {
-				climateReactState.lowTemperatureState.swing = 'both'
-				climateReactState.highTemperatureState.swing = 'both'
-			} else if (state.verticalSwing === 'SWING_ENABLED') {
-				climateReactState.lowTemperatureState.swing = 'rangeFull'
-				climateReactState.highTemperatureState.swing = 'rangeFull'
-			} else if (state.horizontalSwing === 'SWING_ENABLED') {
-				climateReactState.lowTemperatureState.swing = 'horizontal'
-				climateReactState.highTemperatureState.swing = 'horizontal'
-			} else {
-				climateReactState.lowTemperatureState.swing = 'stopped'
-				climateReactState.highTemperatureState.swing = 'stopped'
-			}
-		} else {
-			if ('verticalSwing' in device.capabilities[state.mode]) {
-				climateReactState.lowTemperatureState.swing = state.verticalSwing === 'SWING_ENABLED' ? 'rangeFull' : 'stopped'
-				climateReactState.highTemperatureState.swing = state.verticalSwing === 'SWING_ENABLED' ? 'rangeFull' : 'stopped'
-			}
-
-			if ('horizontalSwing' in device.capabilities[state.mode]) {
-				climateReactState.lowTemperatureState.horizontalSwing = state.horizontalSwing === 'SWING_ENABLED' ? 'rangeFull' : 'stopped'
-				climateReactState.highTemperatureState.horizontalSwing = state.horizontalSwing === 'SWING_ENABLED' ? 'rangeFull' : 'stopped'
-			}
-		}
-
-		return climateReactState
 	}
+
 }
